// src/types/pokemon.ts
export interface SearchFilters {
  query: string;
  type: string;
  sortBy: 'id' | 'name' | 'height' | 'weight';
  sortOrder: 'asc' | 'desc';
  showFavorites: boolean;
}